The following changes have been made: 1) restructured the code on the server to create a proper flask app 2) started adding input formatting and exception handling 3) started reworking the functions to expect input in the body of the request
